linear analyses with multiple observables per an instant
